move ReservedWord enum to ExternalTool #3657
